Add class for Nino3.4 index
