+++ mpas_analysis/ocean/__init__.py
@@ -2,5 +2,5 @@ from climatology_map import ClimatologyMapSST, ClimatologyMapMLD, \
     ClimatologyMapSSS
 from time_series_ohc import TimeSeriesOHC
 from time_series_sst import TimeSeriesSST
-# from nino34_index import nino34_index
+from index_nino34 import IndexNino34
 # from meridional_overturning_circulation import moc_streamfunction
